Kernel/Boot: Start of a kernel
This might be a lot in one commit, took quite some time to get it
finished :D, so here are some notes:

So what this is currently is a multiboot ELF file that is bootable by
`qemu`'s bootloader (with `-kernel` flag) and possibly other bootloaders
such as `GRUB`

Normally, we can setup a multiboot header with just the `magic` and
`flag=0`, but then it won't work (at least qemu doesn't boot it) because
it expects ELF32 file and here we have ELF64.

That's why we need to set the `flag=0x10000`, TBH I'm not sure what's
this flag stand for, but if this flag is provided, we have to provide
the extra loading pointers data after the `checksum`.

All of the data is generated from the linker, since the linker has
access to all the pointers calculations.

We are using this form `load_addr + (<something> - begin)`
where:
- `begin` is the start of the virtual address (where the kernel will be
  mapped to later on).
- `load_addr` is the start of the physical address.
- `<something>` will be an address in the virtual space.
So, this basically convert a virtual address to physical address by
using the offset from `begin` and `load_addr`.

Of course after boot, in `entry` is running in 32bit protected mode, but
all the code compiled by `rustc` will be in `x64` long mode, and thus,
we can't start executing rust code, that's why the `entry` is in
assembly (will be moved later to a separate file).

On boot we are checking that the `eax` has the correct value `0x2BADB002`,
which means that we are booted by multiboot and recognized.

So, the next steps is to setup long mode and jump to rust code from it,
then we can continue the happy path.

This is maybe the most compact and nice way I found that it works, not
sure if this method is better.
The other method is setting up a hard drive and loading 2
components, the first stage which is 32bit, and then it will read the
file system (probably FAT), extract the 64bit ELF file, load it in
memory and execute it.

I also don't want to use any external library, at least for now, that's
why I didn't use `bootloader` or `bootimage` or whatever.

-- The journey begins