Kernel/Console: started implementation and added normal `print/ln!`
Now finally back to normal rust prints.

This is a big commit, contain a lot of stuff, and a lot of issues as
well.

Now, we have the Console. Inside it, our old friend VgaBuffer, which
prints to the VGA screen.
We also added this time, the UART, the setup currently uses IRQ and
FIFOs, which we are not utilizing now. We should do in the future.

But beside that, the UART works in a synced fashion.

The issue now is how we should handle locking and panics with the console.

In a multi-processor case (which we are reaching), several cases might
happen:

1) Cpu1 panicked, but is not in the console code (console is not locked
   by us)
    - If the console is not locked at all, we can just grab the lock and
      never release, and get stuck in an infinite loop after printing
      the error.
    - If the console is locked by other Cpu, we can wait for the lock to
      be released and do the above.
2) Cpu1 panicked, and we are in the console code (console is locked by
   us)
    - We can just print (WITHOUT taking the lock), since its already
      locked, and then just wait in an infinite loop.

These might be not the best solution, but I'll see how rust implement
`stdio` and do something similar.

Anyway, we have console. Since we are still in 1 cpu land, no issues.