Kernel/Sync: created `Mutex` and `ReMutex`
Now we have more mature sync primitives now, which allows for easier
and safer usage of structs.

There are some issues/downsides still:
- Because the lock is around the whole thing, it creates some
  performance impact, if we look at the physical_page_allocator.
  After this commit, when we allocate some page it will do the following
  while holding the lock, which will make other CPUs wait for more time:
    - get the page.
    - fill it with garbage (not needed to perform while locked).

Regarding the issue from last commit, on the console lock,
Now with using RefCell and ReMutex, I think its a bit better.
But if for example Console::write_byte is the one causing panic, then it
might just result in an infinite recursion.
But this is very unlikely, since we are creating a new `Console` object
and re-initializing the VGA and UART communication.