Kernel: use `hint::spin_loop`
Replaced calling `pause` with this, even though its a bit worst.

The reason I wasn't using `_mm_pause` function is that its not inlined,
Looking at the assembly, it produce a call to `_mm_pause` function,
which just executes `pause; ret`.
After thinking about it, I think its not that important, since we are
waiting anyways, and this will make it generic over other architectures
if I decided to port this into.