Kernel/VirtualMem: Added virtual memory management, and remapped the kernel
Create a virtual memory manager, and reload new pages.

This manager, will map memories from a ptr in physical to a ptr in
virtual with length and flags.

The code is very specific to 64-bit x86 architecture, and will need to
be changed if ported to other architectures.

The kernel is now mapped to the virtual address 0xFFFFFFFF80000000, we have 3 mapped sections:
- Low memory (has some BIOS stuff)
- Extended memory: kernel .text and .rodata sections
- Extended memory: kernel .data and .bss sections and the rest of the data for the `whole` memory
  we decided to use in the kernel.
  And this whole memory is 128MB for now starting from 0xFFFFFFFF80000000.