Kernel/Heap: Created `global_allocator` and added `alloc` crate
Now we can use `Vec`, `String` and all the wonderful things.

Took some time to implement this allocator :D, but I guess its working
now. Did some heavy testing on it.

We should have some sort of unit testing for these components, such as
the allocator and so on.

Currently, there is a specific size on how much we can use (16MB), this
is not new memory, but is part of the (128MB) that is owned by the
kernel.

The virtual address of the heap starts after the 128MB mark.
This is not mapped directly to physical space. So we are not writing
over any of the IO memory-mapped registers there.