Kernel/Interrupts: added IDT and interrupts
This is building from the previous one (GDT), and is quite important.

Now hopefully, the QEMU won't just restart when we get an exception but instead we will handle it.

There is something I'm not sure about regarding this.
I tried enabling interrupts with `sti`, but I get `double-fault`(8).
This is one issue, the other issue is that the data being printed is not correct, it seems that this handler shouldn't have `Error code`,
but when looking at the specs, looks like it should.

I'm not sure if what's being triggered is really page-fault. very strange.

Another issue, is that I tried causing `stack-overflow` exception,
but for some reason, qemu restarts without calling any of the exception handlers.
This is very strange, as, I thought this should handle everything.

At least we have this now, it could be there is a bug somewhere, I'll fix it later.