Kernel/IDT: fixed issues with infinite loop interrupts
Turned out I was not using the TSS stacks, since I was using entry `0`, which means,
keep the same execution stack.

So, I fixed the API, so that we can specify that we are adding a special stack index.

I applied this already to default handlers that deal with faults,
such as page-fault and double-fault.

This fixed the issues where stackoverflow crashes qemu. it was triggering an exception
while trying to handle the page-fault exception, but using the same stack.

I'm still having the issue where `set_interrupts` trigger double-fault without error-code (the enteries in the stack are not correct).