Kernel/Memory: Mapped memory-mapped regs into virtual memory
These are provided by the bios and other hardware (CPU).

The bios structures, are a bit over the place, there are some before the
middle ram (after 0x10_0000), and there are some after that, which we
don't map.

So added mapping for them.

Beside that, fixed some bugs related to the multiboot_info still using
physical addresses, which results in crashes when printing the content
after mapping virtual memory.

And used the format 0x0000_0000 (underscores) where applicable for const
variables, since its much easier to read the values like this.