Kernel/Heap: fixed a bug where we corrupt allocated blocks sometimes
This happen when you allocate almost all of a `free` block, and when we shrink it to create a new `free` block,
we overflow the space that we have.

An example of how it looks:

- Here is the current free state
000..100: Free
100..200: Used

If our `HeapFreeBlock` lets say is `10` bytes for example.
If we allocated 95 bytes (including header and so on).
What will happen is that we will do the following
000..095: Used
095..100: Free
100..200: Used

And now we have a free block that is `5` bytes long, but the data it needs
to store the metadata about itself is `10`, so we have corrupted
the entry at `100..200`, and when it gets freed, the magic value
will be corrupted.

It was a nice idea to add this magic value otherwise idk how I would have
tracked this haha.

The fix for this issue is to check that the new empty is capable of storing the
needed space for at least an empty block.

If that happened, we would have a literally useless block, but meh,
that's fine.

Since the size of the block now might be larger than the layout
requirements, we adjusted `dealloc` to accommodate that.