Kernel/Mutex: Add `run)with` to run callback on the locked value
This is useful as it allows us to get the lifetime reference of the inner value
to be the same as the outer value.

This is useful for when we need a static lifetime inside a lock.
Its technically possible since the outer value is static, then the inner must also be static.

But using the normal method of locking and using `Deref` doesn't allow that
So, that's why this is added instead.