Kernel/APIC: setup APIC MMIO struct and definitions
Its annoying dealing with this, since its an MMIO, you must do a 32bit write.
but the size of each entry is quite large (16 bytes), so that's why we have this weird structure.

Another method to do this is to add padding everywhere, but it felt very annoying.

So I settled with this.