Kernel: some cleanup, removed `static mut` when not needed
Its still used in the heap and physical allocator, because what `static mut` does
is that its like `unsafe`, it tells the compiler to not do safety checks.

If I removed `mut`, it won't compile, even though its using `Mutex`,
because these allocators contain raw pointers, which is not `Send`.

Meaning, that if for example I created another allocator, they will share the same memory values addresses
which will create major issue.

Of course, I'm not creating 2 allocators, but I thing keeping `unsafe` is better
so that looking at the code, it will be clear that this part is dangrous and should be handled with care.