Kernel/Console: Added keyboard implementation and interrupts
A lot in this one.

Added a keyboard driver (just mapping scan_codes).

Added interrupts for keyboard and uart using IOAPIC.

Currently, the console holds the keyboard.
This is not ideal, as we want the keyboard to be standalone entity.
So that we can get access to the scancode information for later when we implement GUI.