Kernel/IDE: identified all drives/devices in a PCI mass storage device
Each device has a primary and a secondary ports, and each of these
support 2 devices, so in total we might have 4.

Currently, we are all storing them inside 1 PCIDevice, since
`probe_init` return 1 device. Maybe would be good to return an iterator
so that we can have multiple devices returned from a single PCI device.
This might make it easier to interact with devices.