Kernel/Heap: Fixed deallocation bug, cycle is created in the free list
// we are not sorting before this commit
The issue was if the free blocks were like so:
- 6000..7000
- 1000..2000
- 8000..9999

If we freed `3000..4000` it will be added to the top:
- 3000..4000
- 6000..7000
- 1000..2000
- 8000..9999

Then if you freed `2000..3000`

The result will become:
- 1000..4000
- 6000..7000
- 1000..4000
- 6000..7000
...

And will result in an infinite loop.

Why is that happening?
Because when freeing `2000..3000` we notice that its between two blocks,
`1000..2000` and `3000..4000`, and thus we can create a new block
`1000..4000` and assign the previous of `1000` to be the previous of
this new block, and the next of `4000` to be the next of this new
block...

The previous of `1000` is `6000..7000` and the next of `4000` is the
same block `6000..7000`. So previous and next are the same block, which
result in this cycle.

The issue can easily be fixed by making sure we always free in order, we
always place freed blocks in the correct order.

I also added some debugging, so that we will detect when a cycle happens
or when an infinite loop happen. (we can disable when we want
performance).