Kernel/FileSystem: started FAT filesystem implementation
I didn't commit for a while, so this is quite a large one.

Here we have implemented FAT12/16/32.
we can list directories and attributes of the files.

Also added `FILESYSTEM_MAPPING` which maps a specific entry to a filesystem.
This will allow us later to map other filesystems inside some directories.