Kernel/VirtualMem: Improved the vm to allow allocation
We can now allocate, without the need to provide a specific physical
address. This is the most useful type, as most of the time, we don't
care where this is being loaded.

Because we want to not provide physical_address, we have to use `size`
and not `physical_address_end`. So this was changed as well.