Kernel/VirtualMem: improved vm and prepared for user_space paging
Added functions to allow cloning a VM
We are using the same L4 entry as is, thus, if we allocated a new page
from another thread, it will be visible to all instances of the kernel,
and its easier to clone this way, no need to remap everything.

Also, user vms can only map user pages, currently we are using `assert`,
but it would be better to return an error later on.

This will protect from malicious ELF files trying to be mapped into
kernel memory.