Kernel/Interrupts: added a new type of interrupt handler that saves all regs
Implemented from assembly without using llvm conventions, we are saving
all registers, and we have access to `vector_number` easily.

This is very nice, but I kept the old ones, since this may be a bit
slower with all the overhead used in saving registers.
But its a lot more robust than what we had.

Now, we can continue working on scheduling easily.