Kernel/Mutex: Improving the locks
Added some traits implementations for mutex
This matches the std Mutex features.

We don't need to make the raw lock `unsafe`, since its fine to call
`lock` multiple times, and that's the whole idea :D, we just want to
make sure same CPU doesn't do that in Mutex, but its fine in ReMutex, so in the eye
of `Lock`, it should be fine, and its not considered undefined behavior.