Kernel/HPET: added HPET timer
Currently, we are initializing the timer and also setting up interrupts,
even though, this may not be needed.

Looking around, looks like this should not be used a primary timer, but
as a source of truth, where `TSC (Time Stamp Counter)` can be calibrated
with this and we just use TSC, and re-calibrate one in a while.

But for now will keep it and remove it later