Kernel/RTC: added RTC date parsing, now we can know what time is it
Currently its not being used except for logging, but at least now,
timekeeping is easy, we have a base of the clock from RTC, and we will
use HPET/TSC to keep the time going from that point.

But not sure how to sync exactly, as RTC getting the data isn't the
fastest thing, though its 1 second difference so its quite a lot.

Let's see