Kernel/Console: create a late console that acts as a device
This is not very nice, I have to say. I'm not sure how to easily switch
the early console to a new console (we are planning to add more feature,
so it won't look the same at all).
Even with this, it will perform Enum checks until the end of time even
though we know `EarlyConsole` will never be back.

:(

Anyway, lets think of what to do with this later, at least we get
something going.

So the plan for this console is to be a device, that we can easily open
a file descriptor to it so that we can use with user mode easily.

I got rid of the whole keyboard and UART interrupts, so now, we can't
get any input events from the user.

An idea I have is to have it lazily read from another `keyboard` device,
and we will have a reference to the keyboard device at all time.
Maybe I'll have a UART device also and make it `read/write` easily.

This may be "after next" plans, now the goal is just to get something
printed in the usermode. So lets ignore all of these improvements for
now and focus on them later.