Kernel/Syscalls: started working on syscalls
Added a new syscall interrupt (0xFE), that will switch user -> kernel

Currently, we have an empty `sys_open` function that should open files
and return a file descriptor.

This is not good now, we need to create a common crate that we can use
in the kernel and usermode stdlib.