Kernel/VirtualMem: fixed mapping unaligned addresses
We were taking the `size` and aligning it alone.
This was causing issues, where last page would not be mapped.
Example:

Mapping:
- virt_addr: 10, size: PAGE_4K

This would result after processing and aligning to
- virt_addr: 0 (aligned down), size: PAGE_4K (aligned up)

And here is the issue, address at `PAGE_4K+1` for example, won't be
mapped but it should.

So we changed it so that we use the `virtual_end` to calculate the size,
and the above mapping will become:
- virt_addr: 0 (aligned down), size: 2*PAGE_4K (aligned up after
  calculating `virtual_end`)