Kernel/VirtualMem: dynamically initialize physical pages, and map bios memory
This will help us be more dynamic, and we don't have to be strict on the rules of at least 128MB.

Also having the bios memory not mapped, we can use that space for something else.
For now its not being used, but maybe in the future.

This will bring us one step closter to using UEFI