Kernel/ACPI: Add StateNamesList tracking current names in AML parser
There was a bug from the previous code that fixes the unknown methods.

An issue happened because looks like, some method invocation can have
another name as argument, so we have to distinguish if the argument is a
name or another method invocation.
For this, we look for names we have seen, from what I saw, I think you
can't have a named object/field and a method sharing the same name.
(hopefully I'm not wrong :D).

This of course isn't fault proof, because it depend on the code after
the call, so if it can fit inside a method invocation our prediction
might be wrong.

But yah, this doesn't need to be super accurate, the information we need
should just be available inside a Package.