Kernel/ACPI: improve AML parsing performance
Get rid of `scopes` as its a very heavy and complex structure to update.
And instead we have done some optimizations to find a better way to
`find_method`.
The changes in design done in this commit are not fault-proof, so we sometimes might get wrong number of arguments,
these are just assumptions, estimations and predictions.
But that's fine, since the correctness of parsing is not really that
important, since the method itself knows how many arguments it uses and
that's good enough. But we try to do correct parsing anyway.

Changes:
- Removed `scopes` recursive structure and keep `methods` and `names`, which are global and shared by all scopes.
    - This assumes that duplicate methods (found in different Devices) have the same arguments since they are standard/common methods.
    - So now, when we want to find a method, we take the last 4 characters and strip all the scope prefixes, and search for it in `methods`. (that's why the assumption above is important)
- Rely on `predict_possible_args`, and improved it further to break on any expression that is not valid or shouldn't be
  a method argument.
- Remove `Rc` and `RefCell`, and share the same `&mut`, since there is only 1 parser active at a time, no need to bloated it
  with `Rc` and `RefCell`.