Kernel/VirtualSpace: added dynamic `virtual_space` allocator
This is for objects and stuff that we know where they are physically,
and we want to use them, but we don't have a specific space we care
about virtually.

And this is to fix #7.

So this will allocate pages from the virtual space so that we can map
them together with the physical address that we know about.

We can do the following:
- get_address: try to find virtual mapping of a physical address if we
  have it, this will panic if you tell it you want some previous
  allocation but larger (check below).
- ensure_size: the address you give it is virtual, so it will find the
  mapping (must exist), and make sure the mapping is enough for your
  `size.
  If its not enough, it will check if it can allocate the space
  immediately after it, and if so, it will do that, otherwise it will
  panic.
- allocate: explicit allocation, this can be called with same physical
  address multiple times and it will result in multiple virtual
  mappings, should be used carefully as it can cause aliasing issues.
- deallocate: de-allocate virtual space.
  This can be probably used after for example reading all ACPI tables.
  At that stage, the memory won't be needed.
  But for other cases, such as memory mapped IO/buffers, it will stay
  throughout the lifespan of the OS.

There is several panics and we used explicit `ensure_size` instead of
making `get_address` do it, I think its better this way to reduce bugs.