Kernel/Memory: added `align_range` to easily align a range of addresses
This is a common pattern we have done before in the `virtual mapper`, and we wanted to add to the `virtual space`. So decided to move
it to an external location.

With this change, we don't have to align the addresses in `acpi::tables`, so that's good and more flexable for us.