Kernel/VirtualMem: fixed user stack was inside kernel region and not cleaned
I thought we were using the whole kernel region and thus just subtracting from it would give us the highest user page.

But this is wrong, as in the kernel, we are only using 2GB from the 512GB accessable from the last L4 entry.

Because of that, we can't just subtract from the beginning of the kernel, user are somewhere else.

With this, we can unmap user stack correctly.