Kernel/Scheduler: added `Yielded` state, to make scheduler more fair
This is not true :D.
So before this commit, the scheduler would do this.
Imagine you have only 2 processes:
p1 and p2.

The timer/syscall interrupt will for example yield from p1, which will set p1 to `Scheduled`, then in the scheduler loop we will take it first, since its the first one and execute it, and so on.
So we will be stuck in the first one.

With this commit, its almost the same, but instead of being 1 process being stuck at, we get 2 :D

So nothing changed, but at least we can test multi processing with 2 processes.

We need to fix this, but this is kinda a quick fix (maybe will remove this state later).

---

Also modified the `exit_current_process` function to use
`with_current_process`.