Kernel/Process: added `heap` management in `Process`
Not tested yet, but we can manage heap now.

The heap location is after the ELF and grows upwards toward the stack.
There will also be dynamic libraries there between the stack and heap also.

This is how linux does it, windows does it differently, but I think this is more flexible and gives more space for the heap.

I'm planning to make something like `sbrk`, compared to using virtual memory.