User+Lib: added `user_std` library and allocator in it
This is quite a large change, since I added the allocator and also a spinlock `Mutex` and `OnceLock` for it to work.
These are taken directly from the kernel and removed `cpu` related stuff since we don't need to worry about interrupts.

But anyway, besides that we got the allocator going in user mode, we can use `format!`, `String`, `Vec` and many more fun stuff.

The next steps I'm planning to see if its possible to close the rust std library and modify it manually step by step until we get something going, then we can publish back as a triplet target.