Kernel/Keyboard: added `keyboard` to the `console` to enable reading from it
This is the first step in getting stdin working in user mode.

Of course, this is the base and should only be used by the first tty, later on, we will use pipes for shells to communicate with child processes.

But we still have issue with this, currently, it will only try to read, and won't spend a lot of time waiting, so if there is nothing there it will just return empty string.
We should implement blocking next.