Kernel+User/File: Added `BlockingMode` to control blocking behavior of files
We want to make it easier to read `stdin` files, thus we have `Line` blocking mode.
The last type of blocking is not used for now, maybe it will be removed in the future or replaced by something else.

But now we should be able to read stdin without issues.

But there is a bug in this and how we handle reading files for now.
It will be fixed in the next commit.