Kernel/File: fixed bug on reading blocking files (getting stuck)
When reading a file that relies on interrupts, and we are reading from a
process we get stuck.
Because, when using the `Process` struct, we take a lock, which disables
interrupts, then when read, we will get stuck since we are in a deadlock
situation, we are waiting for interrupts that will never fire.

This can be fixed technically without the hack, by relying on the
Console reading from the `keyboard` all the time, but this is a general
dump fix.

The best plan is to have a waitable objects so that we can wait for the
file to be ready, this fixes several issues:
- We won't get suck in a deadlock.
- Reduce CPU load, since it will be waiting but not looping, and thus, the
  CPU can do other stuff in the mean time.