Kernel/Process: added support for processes interrupt in kernel
Previously we couldn't interrupt a process while it was running in the
kernel, turned out easily removing that check in the
`yield_current_if_any` doesn't do it :D and introduce a rabbithole of
debugging.

- What was happening?

After enabling interrupts in the kernel, the system would crash in a
weird way.
After a lot of time debugging, here is what was happening.

- Process 0 running in user mode                      (all good)
- Process 0 yields from user mode                     (all good)
- Process 1 running in user mode                      (all good)
- Process 1 goes to kernel mode (syscall for example) (all good)
- Process 1 yields from kernel mode
  (this is not yielding after the syscall like we did
   before this commit, but timer interrupted while in
   kernel mode and yielded, so the state is saved
   of the process while we are in kernel mode)        (all good?)
- Process 0 running in user mode                      (all good)
- Process 0 goes to kernel mode (syscall for example) (BAD) // explained later
- Process 0 goes back to user mode                    (...)
- Process 0 yields from user mode                     (...)
- Process 1 running in kernel mode                    (Catastrophe!)

And here we reach the bad stage. But it seems ok, I don't get what's
happening.

What happened is that `Process 1` is still using the kernel resources,
even though we thought it yielded, and we saved its whole thread context,
but we didn't, the kernel stack used for handling syscalls will be overwritten
by `Process 0` on the `BAD` line. And then, when the `Process 1` takes back
execution in `Catastrophe` line, the whole stack is modified. The state
of the process was not saved correctly, because we didn't account for
the stack.

So, how can this be saved?

Easy, just save the stack as well as the other registers.
This took a while to get right [you only see the correct implementation
in this commit :'( ]

But here is what's happening in this commit:

- Doesn't change much, but we removed the `SYSCALL_STACK`, since we can
  just use the `kernel ring stack` and save up on how many stacks we
  need to save. This `kernel ring stack` is where the stack will be used
  in case we came from the user mode.
  Without this change, we can still fix the issue.
- Setup a per-process kernel stack, when a process moves to the kernel
  by any means, it will use this stack.
  This is saved in the process Kernel region, which is a region in the
  virtual memory for kernel memory associated with each process.
  `FFFF_FF80_0000_0000..FFFF_FFFF_8000_0000`
  This will be switched on VM switch, so that's how we can `save` the
  stack along with registers.
- BUT, because the stack now is part of the VM, we can't just switch
  between VMs easily. So, here are the changes affected by this:
  - `switch_to_kernel` function, since we can't do that inside an
    interrupt handler while we are using the kernel stack, we moved it
    after the scheduler goes to kernel.
  - When loading the `ELF`, we clone the current VM (not the kernel VM),
    so that we keep all the process specific stuff, after we load the
    kernel, we replace those process specific memory with our own new
    memory. That's why we call `add_process_specific_mappings` after
    loading the elf, after this stage, we should never switch to this VM
    unless we use the scheduler.
- Another note, is that since now we can be interrupted in the kernel,
  there is a gap between when the scheduler loop selects a process and
  calling `int FE` to run it, and in this gap, the timer can interrupt
  and yield the process while we think we still have it :D
- Another note, is that since now we can be interrupted in the kernel,
  there is a gap between when the scheduler loop selects a process and
  calling `int FE` to run it, and in this gap, the timer can interrupt
  and yield the process while we think we still have it :D.
  To fix this, we added `scheduling` variable in the `Cpu` to indicate
  that we shouldn't be yielding anything.

  I tried to do it with `push_cli` so that we don't get interrupted, but
  for some reason it was still failing sometimes, maybe there is a gap
  somewhere and it manages to interrupt and steal the process context,
  I think this is a better and more verbose method so will keep it as is
  for now.

And yea, that's the end.
Its funny, several times while debugging this, I said `I got it`, but
another issue comes up, or I didn't get it at all.

I'm not sure if this is enough when multiprocessing comes into view, but
hopefully it won't be a lot of issues :D