Kernel/Pipe+File: Added `device_clone` event for devices and file cloning
Without this we would close the pipe if a clone was closed which is a thing we don't want.

Now this is managed by devices.
But maybe we should move this to be global to all files, and only call `close` once all of them are closed, i.e. instead of being handled by `Pipe` it will be handled by `File` for example