Kernel+Lib: added `open/read_dir` syscalls to read directories
This will return `FileStat` and filename, the filename is only for that
file. The caller must keep track of the parent.

`open_dir` will return an `fd` that can be closed later, but must be
used with `read_dir`, `read` cannot be used and will result in an error
`IsDirectory`.

Beside that, all is basic. Now this is not optimal, as it would be
better to dynamically read the directory from the kernel instead of
reading the whole directory on `open_dir`.