Lib: renamed `DIR_ENTRY_MAX_NAME_LEN` to `MAX_FILENAME_LEN`
This will become more general for later checks to not overflow filename length

Not sure if there is any constrains with ext2/3/4 and other advanced filesystems. But for now I set it like so.

This is not the most efficient, we are wasting all the free space as most files are very small in name.

For now we are doing this as its simpler to have constant `DirFilename` size, we can do more complex encoding later on as part of performance improvement