Lib/Heap: fixed another bug in the heap allocator with alignment
This is an addition to another fix I did recently on the same issue.

Turned out that fix (df5e7c6) didn't fix much :D since we can't just add
`layout.align()` into `possible_next_offset` that would result in
needing more memory than what we need.

Its good that there was that `assert!(allocated_block_offset <= block_offset_from_header)`
which helped catching the issue.

Now, if we exceed the size we just assume we have more memory and go on,
panic if we actually don't.

We can just request more memory if that happen.

Another idea is to just request more memory upfront, so we always
request `allocation_size + layout.align()` all the time, of course we
only use what we need. But this will reduce going back and forth.

In general this approach now is less efficient a bit, since we might use
way more memory than what we need.
But this is generally only applicable for blocks that have more
alignment than 16, which probably is quite rare.