Kernel/Process: added `current_dir` for a `Process`
This controls the current directory context of the process.
We are storing it as `fs::Directory` as this is a valid running
directory (at least for now since we can't delete directories).
This allows us to be sure that we are using a valid directory and reduce
checking.

Currently this is never used anyway so it doesn't really matter, next,
we will add features to allow using relative paths based on this
`current_dir` in the process.