Kernel/Filesystem+Fat: Added treversing support without need for `/`
Since we have added the `Path` and stripping and so on, when we get the `remaining` from the filesystem mapping, we don't get the root character with it.

Previously we had to manually put `/` in the beginning which of course perform allocation, but there is no need with this now.

It still sad that we do treversing in fat, it would be best to just read a directory from inode and traverse ourselves, buts its more efficient now, until we have a better solution for general treversing.