Kernel/Syscall: Added support for relative path for all related syscalls
Added `sys_arg_to_path`, which is similar to `str`, just convert to
`path`.

Then we have added `path_to_proc_absolute_path`, that will try to see if
its a relative path, and if it is, it will use the `current_dir` from the
`Process` along with the argument relative path to get the full
absolute_path.

If it is an absolute path, we just use it directly.

This is because `open_inode` and such functions rely on the path being
absolute.