Kernel/Boot: fix not detecting 64bit correctly for Intel CPUs
In intel, the features in `CPUID 0x80000001` have the `6th` bit reserved and not copied from `CPUID 0x1` like AMD.
This would cause it to fail on intel if the hardware doesn't use these bits (it could, just undefined).

Also removed unused variables in the assembly forgot since moving to multiboot2, and other improvements.