Kernel/ACPI: Improved predicting methods in AML parsing
predicting number of args is one of the annoying issues when parsing AML, any method to improve that would greatly improve the parser.

toward that, instead of having only a specific `parse_term_arg_for_method_arg` that can accept being a method, we reversed that, so that all `TermArg` can have method call, and kept a small number (3) of fields that don't accept that due to a bug in our parsing (should be removed later on).

Along with removing fields that would break the prediction, like logical operators, these should work fine as an argument to a method call.

There are several improvement ideas to add as well here.
1- Perform breadth-first parsing, i.e. parse all objects and fields,
then methods in a scope, this will help us a lot since named objects are
referenced inside methods but created after the method, so we do
prediction, which is bad.
2- I notice that inside the prediction we go into very deep prediction
chain that is not possible, like parsing `if-else` expression, which
will never be in an invocation for a term-arg anyway, so we should quit
searching if we reach similar stages.