Kernel/Boot: more stack size (2MB)
The only thing that take a lot of stack is parsing AML, we should change
it so that it stores data more in the heap and not in the stack.

Reorganizing the code to be breadth-first parsing aligns as well with
this issue, as it would use less stack space.

For now we can add, 2MB is okay