Kernel/VirtualMem: fixed bug add overflow on physical address above `KERNEL_END`
So the issue is like this. When we map physical to virtual in the `virtual_memory_mapper`, i.e. we add it to the page table, we store the physical address in the page.
But when we get the page back using `PageDirectoryTablePtr`, we auto convert it in `from_entry` to virtual.

A lot of stuff can still be improved here, but one step at a time :)

The issue is that conversion is simple `+ KERNEL_BASE` which assumes that the physical address is in the kernel low range.

But this is not true for all cases, when we map physical ranges above that this will trigger add-overflow bug, and this is not even useful, we know that since we have used `virtual-space`, this address won't be freed for example in the `physical_page_allocator`, so there is no need to convert it to virtual address.

This is basically what this commit is for, we store physical address in the `PageDirectoryTablePtr` and only convert it to virtual when needed along with `assert` to make sure we are not going to trigger a bug.